+++ CHANGES.md
@@ -11,6 +11,10 @@
 
 ## develop
 
+- [ADD] 受信している接続のクライアント ID の表示に対応する
+  - `notify` で受け取ったクライアント ID を表示に使用するため、state の `soraContents.remoteMediaStream` を `soraContents.remoteClient` に変更し、MediaStream の他に `connectionId` と `client_id` を保持できる型に変更する
+  - この変更に伴ってリモートの `MediaStream` を使用した関数、変数の名前を `Client` に変更する
+  - @tnamao
 - [CHANGE] `Session ID` と自身の  `Connection ID` `Client ID` の表示を `type: notify` の `connection.created` を受け取ったタイミングでの表示に変更する
   - この変更に伴い、Sora Devtools の Sora 接続状態の確認は state の `soraContents.connectionStatus` の値の確認も追加する
   - @tnamao